Add files via upload
A small robot-run cafe in Los Angeles is in the works. The project is promising but expensive, so the company plans to attract investors. They’re interested in the current market conditions—will Robo Roasters maintain success when the novelty of robot waiters wears off? My task is to prepare market research with open-source data on restaurants in LA.